Fix build errors with recent boost libraries (>= 1.67)
With https://github.com/boostorg/date_time/commit/f9f2aaf5216c
we need to update the usage of duration.
Reference: https://github.com/Icinga/icinga2/pull/6230

+++ rostime/include/ros/impl/time.h
@@ -165,9 +165,9 @@ namespace ros
   {
     namespace pt = boost::posix_time;
 #if defined(BOOST_DATE_TIME_HAS_NANOSECONDS)
-    return pt::from_time_t(sec) + pt::nanoseconds(nsec);
+    return pt::from_time_t(sec) + pt::nanoseconds(long(nsec));
 #else
-    return pt::from_time_t(sec) + pt::microseconds(nsec/1000.0);
+    return pt::from_time_t(sec) + pt::microseconds(long(nsec/1000.0));
 #endif
   }
